+++ morphman/common/__init__.py
@@ -1,6 +1,8 @@
+from .centerline_operations import *
 from .common import *
-from .morphman_operations import *
+from .surface_operations import *
 from .vessel_reconstruction_tools import *
 from .vmtk_wrapper import *
 from .vmtkpointselector import *
+from .voronoi_operations import *
 from .vtk_wrapper import *
