+++ morphman/manipulate_bifurcation.py
@@ -9,7 +9,7 @@ from argparse import ArgumentParser, RawDescriptionHelpFormatter
 
 # Local import
 from morphman.common.argparse_common import *
-from morphman.common.morphman_operations import *
+from morphman.common.surface_operations import *
 from morphman.common.vessel_reconstruction_tools import *
